Localize SearchResultViewController
