+++ DuplicateFinder/AppDelegate.swift
@@ -1,7 +1,4 @@
 //
-//  AppDelegate.swift
-//  DuplicateFinder
-//
 //  Created by Nixon Shih on 2020/10/2.
 //  Copyright © 2020 Nixon Shih. All rights reserved.
 //
@@ -9,18 +6,5 @@
 import Cocoa
 
 @NSApplicationMain
-class AppDelegate: NSObject, NSApplicationDelegate {
-
-
-
-    func applicationDidFinishLaunching(_ aNotification: Notification) {
-        // Insert code here to initialize your application
-    }
-
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
-    }
-
-
-}
+internal final class AppDelegate: NSObject, NSApplicationDelegate { }
