+++ DuplicateFinder/Base.lproj/Main.storyboard
@@ -112,10 +112,10 @@
             </objects>
             <point key="canvasLocation" x="42" y="272"/>
         </scene>
-        <!--Select Directory View Controller-->
+        <!--Search Settings View Controller-->
         <scene sceneID="hIz-AP-VOD">
             <objects>
-                <viewController id="XfG-lQ-9wD" customClass="SelectDirectoryViewController" customModule="DuplicateFinder" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="XfG-lQ-9wD" customClass="SearchSettingsViewController" customModule="DuplicateFinder" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" id="m2S-Jp-Qdl">
                         <rect key="frame" x="0.0" y="0.0" width="510" height="437"/>
                         <autoresizingMask key="autoresizingMask"/>
@@ -408,10 +408,10 @@
             </objects>
             <point key="canvasLocation" x="-43" y="790.5"/>
         </scene>
-        <!--Logo Info View Controller-->
+        <!--Welcome View Controller-->
         <scene sceneID="XE4-9r-jX4">
             <objects>
-                <viewController id="Me5-Ig-8Zm" customClass="LogoInfoViewController" customModule="DuplicateFinder" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="Me5-Ig-8Zm" customClass="WelcomeViewController" customModule="DuplicateFinder" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" id="RBd-p6-DWW">
                         <rect key="frame" x="0.0" y="0.0" width="190" height="300"/>
                         <autoresizingMask key="autoresizingMask"/>
@@ -467,7 +467,7 @@
         <scene sceneID="77x-W1-YK5">
             <objects>
                 <windowController storyboardIdentifier="SearchFileNameResultWindowSID" id="s6H-VD-NpC" sceneMemberID="viewController">
-                    <window key="window" title="搜尋結果" allowsToolTipsWhenApplicationIsInactive="NO" autorecalculatesKeyViewLoop="NO" releasedWhenClosed="NO" visibleAtLaunch="NO" frameAutosaveName="" animationBehavior="default" id="Y6I-ij-MdO">
+                    <window key="window" title="search_result_window_title" allowsToolTipsWhenApplicationIsInactive="NO" autorecalculatesKeyViewLoop="NO" releasedWhenClosed="NO" visibleAtLaunch="NO" frameAutosaveName="" animationBehavior="default" id="Y6I-ij-MdO">
                         <windowStyleMask key="styleMask" titled="YES" closable="YES" miniaturizable="YES" resizable="YES"/>
                         <windowPositionMask key="initialPositionMask" leftStrut="YES" rightStrut="YES" topStrut="YES" bottomStrut="YES"/>
                         <rect key="contentRect" x="294" y="313" width="480" height="270"/>
@@ -484,10 +484,10 @@
             </objects>
             <point key="canvasLocation" x="807" y="241"/>
         </scene>
-        <!--Search File Name Result View Controller-->
+        <!--Search Result View Controller-->
         <scene sceneID="89e-Ut-xAm">
             <objects>
-                <viewController id="rZF-QJ-B0D" customClass="SearchFileNameResultViewController" customModule="DuplicateFinder" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="rZF-QJ-B0D" customClass="SearchResultViewController" customModule="DuplicateFinder" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" id="6Fw-Cu-BO1">
                         <rect key="frame" x="0.0" y="0.0" width="480" height="276"/>
                         <autoresizingMask key="autoresizingMask"/>
