+++ DuplicateFinder/SearchSettingsViewController.swift
